Make add a float function for now

+++ src/synth/test/add.c
@@ -1,4 +1,4 @@
-int add(int x, int y)
+float add(float x, float y)
 {
   return x + y;
 }
